Revert "Sayfalar Page adı altındaki klasöre taşındı"
