--- lib/Pages/anasayfa.dart
@@ -1,155 +1,0 @@
-import 'package:enter/Pages/main.dart';
-import 'package:flutter/material.dart';
-
-class AnaSayfa extends StatefulWidget {
-  @override
-  _AnaSayfaState createState() => _AnaSayfaState();
-}
-
-class _AnaSayfaState extends State<AnaSayfa> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        color: Colors.grey[800],
-        child: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Container(
-                //Card(
-                //elevation: 4.0,
-                //color: Colors.white,
-                //margin: EdgeInsets.only(left: 20,right: 20),
-                //shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-                child: Padding(
-                  padding: const EdgeInsets.all(10),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      Container(
-                        child: Image(
-                            image: NetworkImage(
-                                'https://www.turizmguncel.com/sites/turizmguncel/uploads/contents/55185/fe0e8d6e690501d2c42fba875eba2aca.jpeg')),
-                      ),
-                      TextFormField(
-                        style: TextStyle(color: Color(0xFF000000)),
-                        cursorColor: Color(0xFF9b9b9b),
-                        keyboardType: TextInputType.text,
-                        obscureText: false,
-                        decoration: InputDecoration(
-                            prefixIcon: Icon(
-                              Icons.email,
-                              color: Colors.grey,
-                            ),
-                            hintText: 'Malzeme Adı',
-                            hintStyle: TextStyle(
-                              color: Color(0xFF9b9b9b),
-                              fontSize: 15,
-                              fontWeight: FontWeight.normal,
-                            )),
-                      ),
-                      TextFormField(
-                        style: TextStyle(color: Color(0xFF000000)),
-                        cursorColor: Color(0xFF9b9b9b),
-                        keyboardType: TextInputType.text,
-                        obscureText: false,
-                        decoration: InputDecoration(
-                            prefixIcon: Icon(
-                              Icons.email,
-                              color: Colors.grey,
-                            ),
-                            hintText: 'Alınan Yer',
-                            hintStyle: TextStyle(
-                              color: Color(0xFF9b9b9b),
-                              fontSize: 15,
-                              fontWeight: FontWeight.normal,
-                            )),
-                      ),
-                      TextFormField(
-                        style: TextStyle(color: Color(0xFF000000)),
-                        cursorColor: Color(0xFF9b9b9b),
-                        keyboardType: TextInputType.text,
-                        obscureText: false,
-                        decoration: InputDecoration(
-                            prefixIcon: Icon(
-                              Icons.email,
-                              color: Colors.grey,
-                            ),
-                            hintText: 'SERİ NO',
-                            hintStyle: TextStyle(
-                              color: Color(0xFF9b9b9b),
-                              fontSize: 15,
-                              fontWeight: FontWeight.normal,
-                            )),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(10),
-                        child: FlatButton(
-                          onPressed: () {},
-                          child: Padding(
-                            padding: EdgeInsets.only(
-                              top: 8,
-                              bottom: 8,
-                              left: 10,
-                              right: 10,
-                            ),
-                            child: Text(
-                              'QR KOD ÜRET',
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 15,
-                                decoration: TextDecoration.none,
-                                fontWeight: FontWeight.normal,
-                              ),
-                            ),
-                          ),
-                          color: Colors.blue[500],
-                          shape: new RoundedRectangleBorder(
-                              borderRadius: new BorderRadius.circular(10)),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(10),
-                        child: FlatButton(
-                          onPressed: () {
-                            Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) => login()));
-                          },
-                          child: Padding(
-                            padding: EdgeInsets.only(
-                              top: 8,
-                              bottom: 8,
-                              left: 10,
-                              right: 10,
-                            ),
-                            child: Text(
-                              'Anasayfaya Dön',
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 15,
-                                decoration: TextDecoration.none,
-                                fontWeight: FontWeight.normal,
-                              ),
-                            ),
-                          ),
-                          color: Colors.blue[500],
-                          shape: new RoundedRectangleBorder(
-                              borderRadius: new BorderRadius.circular(10)),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              //)
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
