+++ lib/anasayfa.dart
@@ -1,19 +1,13 @@
-import 'package:enter/Pages/%C3%BCyeol.dart';
-import 'package:enter/Pages/anasayfa.dart';
+import 'package:enter/main.dart';
 import 'package:flutter/material.dart';
 
-void main() {
-  runApp(MaterialApp(
-    home: login(),
-  ));
-}
+class AnaSayfa extends StatefulWidget {
 
-class login extends StatefulWidget {
   @override
-  _loginState createState() => _loginState();
+  _AnaSayfaState createState() => _AnaSayfaState();
 }
 
-class _loginState extends State<login> {
+class _AnaSayfaState extends State<AnaSayfa> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -23,17 +17,20 @@ class _loginState extends State<login> {
           padding: const EdgeInsets.all(8.0),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Container(
+            children: <Widget>[Container(
+            
+              //Card(
+                //elevation: 4.0,
+                //color: Colors.white,
+                //margin: EdgeInsets.only(left: 20,right: 20),
+                //shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
                 child: Padding(
                   padding: const EdgeInsets.all(10),
                   child: Column(
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: <Widget>[
                       Container(
-                        child: Image(
-                            image: NetworkImage(
-                                'https://www.turizmguncel.com/sites/turizmguncel/uploads/contents/55185/fe0e8d6e690501d2c42fba875eba2aca.jpeg')),
+                        child: Image(image: NetworkImage('https://www.turizmguncel.com/sites/turizmguncel/uploads/contents/55185/fe0e8d6e690501d2c42fba875eba2aca.jpeg')),
                       ),
                       TextFormField(
                         style: TextStyle(color: Color(0xFF000000)),
@@ -44,40 +41,59 @@ class _loginState extends State<login> {
                             prefixIcon: Icon(
                               Icons.email,
                               color: Colors.grey,
+
                             ),
-                            hintText: 'E-posta',
+                            hintText: 'Malzeme Adı',
                             hintStyle: TextStyle(
                               color: Color(0xFF9b9b9b),
                               fontSize: 15,
                               fontWeight: FontWeight.normal,
-                            )),
+                            )
+                        ),
                       ),
                       TextFormField(
                         style: TextStyle(color: Color(0xFF000000)),
                         cursorColor: Color(0xFF9b9b9b),
                         keyboardType: TextInputType.text,
-                        obscureText: true,
+                        obscureText: false,
                         decoration: InputDecoration(
                             prefixIcon: Icon(
-                              Icons.vpn_key,
+                              Icons.email,
                               color: Colors.grey,
+
                             ),
-                            hintText: 'Parola',
+                            hintText: 'Alınan Yer',
                             hintStyle: TextStyle(
                               color: Color(0xFF9b9b9b),
                               fontSize: 15,
                               fontWeight: FontWeight.normal,
-                            )),
+                            )
+                        ),
                       ),
+                      TextFormField(
+                        style: TextStyle(color: Color(0xFF000000)),
+                        cursorColor: Color(0xFF9b9b9b),
+                        keyboardType: TextInputType.text,
+                        obscureText: false,
+                        decoration: InputDecoration(
+                            prefixIcon: Icon(
+                              Icons.email,
+                              color: Colors.grey,
+
+                            ),
+                            hintText: 'SERİ NO',
+                            hintStyle: TextStyle(
+                              color: Color(0xFF9b9b9b),
+                              fontSize: 15,
+                              fontWeight: FontWeight.normal,
+                            )
+                        ),
+                      ),
+
                       Padding(
                         padding: const EdgeInsets.all(10),
                         child: FlatButton(
-                          onPressed: () {
-                            Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) => AnaSayfa()));
-                          },
+                          onPressed: (){},
                           child: Padding(
                             padding: EdgeInsets.only(
                               top: 8,
@@ -86,7 +102,7 @@ class _loginState extends State<login> {
                               right: 10,
                             ),
                             child: Text(
-                              'Giriş Yap',
+                              'QR KOD ÜRET',
                               style: TextStyle(
                                 color: Colors.white,
                                 fontSize: 15,
@@ -96,52 +112,41 @@ class _loginState extends State<login> {
                             ),
                           ),
                           color: Colors.blue[500],
-                          shape: new RoundedRectangleBorder(
-                              borderRadius: new BorderRadius.circular(10)),
+                          shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(10)),
                         ),
                       ),
                       Padding(
-                        padding: const EdgeInsets.all(10),
-                        child: FlatButton(
-                          onPressed: () {
-                            Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) => UyeOl()));
-                          },
-                          child: Padding(
-                            padding: EdgeInsets.only(
-                              top: 8,
-                              bottom: 8,
-                              left: 10,
-                              right: 10,
-                            ),
-                            child: Text(
-                              'Üye Ol',
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 15,
-                                decoration: TextDecoration.none,
-                                fontWeight: FontWeight.normal,
-                              ),
+                      padding: const EdgeInsets.all(10),
+                      child: FlatButton(
+                        onPressed: (){
+                          Navigator.push(context, MaterialPageRoute(builder: (context)=> login()));
+                        },
+                        child: Padding(
+                          padding: EdgeInsets.only(
+                            top: 8,
+                            bottom: 8,
+                            left: 10,
+                            right: 10,
+                          ),
+                          child: Text(
+                            'Anasayfaya Dön',
+                            style: TextStyle(
+                              color: Colors.white,
+                              fontSize: 15,
+                              decoration: TextDecoration.none,
+                              fontWeight: FontWeight.normal,
                             ),
                           ),
-                          color: Colors.blue[500],
-                          shape: new RoundedRectangleBorder(
-                              borderRadius: new BorderRadius.circular(10)),
                         ),
-                      )
+                        color: Colors.blue[500],
+                        shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(10)),
+                      ),
+                    ),
                     ],
                   ),
                 ),
-              )
-              //Card(
-              //elevation: 4.0,
-              //color: Colors.white,
-              //margin: EdgeInsets.only(left: 20,right: 20),
-              //shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-
-              // )
+                ),
+              //)
             ],
           ),
         ),
