+++ lib/üyeol.dart
@@ -1,7 +1,8 @@
-import 'package:enter/Pages/main.dart';
+import 'package:enter/main.dart';
 import 'package:flutter/material.dart';
 
 class UyeOl extends StatefulWidget {
+
   @override
   _UyeOlState createState() => _UyeOlState();
 }
@@ -16,17 +17,14 @@ class _UyeOlState extends State<UyeOl> {
           padding: const EdgeInsets.all(8.0),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Container(
-                child: Padding(
+            children: <Widget>[Container(
+              child: Padding(
                   padding: const EdgeInsets.all(10),
                   child: Column(
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: <Widget>[
                       Container(
-                        child: Image(
-                            image: NetworkImage(
-                                'https://st.depositphotos.com/1038076/5054/i/600/depositphotos_50548309-stock-photo-sign-up.jpg')),
+                        child: Image(image: NetworkImage('https://st.depositphotos.com/1038076/5054/i/600/depositphotos_50548309-stock-photo-sign-up.jpg')),
                       ),
                       TextFormField(
                         style: TextStyle(color: Color(0xFF000000)),
@@ -37,13 +35,15 @@ class _UyeOlState extends State<UyeOl> {
                             prefixIcon: Icon(
                               Icons.pending,
                               color: Colors.grey,
+
                             ),
                             hintText: 'İsminiz',
                             hintStyle: TextStyle(
                               color: Color(0xFF9b9b9b),
                               fontSize: 15,
                               fontWeight: FontWeight.normal,
-                            )),
+                            )
+                        ),
                       ),
                       TextFormField(
                         style: TextStyle(color: Color(0xFF000000)),
@@ -54,13 +54,15 @@ class _UyeOlState extends State<UyeOl> {
                             prefixIcon: Icon(
                               Icons.pending,
                               color: Colors.grey,
+
                             ),
                             hintText: 'Soy isminiz',
                             hintStyle: TextStyle(
                               color: Color(0xFF9b9b9b),
                               fontSize: 15,
                               fontWeight: FontWeight.normal,
-                            )),
+                            )
+                        ),
                       ),
                       TextFormField(
                         style: TextStyle(color: Color(0xFF000000)),
@@ -71,13 +73,15 @@ class _UyeOlState extends State<UyeOl> {
                             prefixIcon: Icon(
                               Icons.pending,
                               color: Colors.grey,
+
                             ),
                             hintText: 'Mail Adresiniz',
                             hintStyle: TextStyle(
                               color: Color(0xFF9b9b9b),
                               fontSize: 15,
                               fontWeight: FontWeight.normal,
-                            )),
+                            )
+                        ),
                       ),
                       TextFormField(
                         style: TextStyle(color: Color(0xFF000000)),
@@ -88,13 +92,15 @@ class _UyeOlState extends State<UyeOl> {
                             prefixIcon: Icon(
                               Icons.pending,
                               color: Colors.grey,
+
                             ),
                             hintText: 'Sicil No',
                             hintStyle: TextStyle(
                               color: Color(0xFF9b9b9b),
                               fontSize: 15,
                               fontWeight: FontWeight.normal,
-                            )),
+                            )
+                        ),
                       ),
                       TextFormField(
                         style: TextStyle(color: Color(0xFF000000)),
@@ -105,13 +111,15 @@ class _UyeOlState extends State<UyeOl> {
                             prefixIcon: Icon(
                               Icons.pending,
                               color: Colors.grey,
+
                             ),
                             hintText: 'Şifre',
                             hintStyle: TextStyle(
                               color: Color(0xFF9b9b9b),
                               fontSize: 15,
                               fontWeight: FontWeight.normal,
-                            )),
+                            )
+                        ),
                       ),
                       TextFormField(
                         style: TextStyle(color: Color(0xFF000000)),
@@ -122,22 +130,21 @@ class _UyeOlState extends State<UyeOl> {
                             prefixIcon: Icon(
                               Icons.pending,
                               color: Colors.grey,
+
                             ),
                             hintText: 'Şifre Tekrarı',
                             hintStyle: TextStyle(
                               color: Color(0xFF9b9b9b),
                               fontSize: 15,
                               fontWeight: FontWeight.normal,
-                            )),
+                            )
+                        ),
                       ),
                       Padding(
                         padding: const EdgeInsets.all(10),
                         child: FlatButton(
-                          onPressed: () {
-                            Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) => login()));
+                          onPressed: (){
+                            Navigator.push(context, MaterialPageRoute(builder: (context)=> login()));
                           },
                           child: Padding(
                             padding: EdgeInsets.only(
@@ -157,20 +164,19 @@ class _UyeOlState extends State<UyeOl> {
                             ),
                           ),
                           color: Colors.blue[500],
-                          shape: new RoundedRectangleBorder(
-                              borderRadius: new BorderRadius.circular(10)),
+                          shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(10)),
                         ),
                       )
                     ],
                   ),
                 ),
-              )
+            )
               //Card(
-              //elevation: 4.0,
-              //color: Colors.white,
-              //margin: EdgeInsets.only(left: 20,right: 20),
-              //shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-
+                //elevation: 4.0,
+                //color: Colors.white,
+                //margin: EdgeInsets.only(left: 20,right: 20),
+                //shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+                
               //)
             ],
           ),
